:recycle: minesweeper.py -> game.py Create Game class
